# src/docs/swagger.yaml
definitions:
...
  models.Category:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
...
info:
  contact: {}
  description: API de controle financeiro pessoal.
  title: SaldoZen API
  version: "1.0"
paths:
  /categories:
    post:
      parameters:
      - description: Dados da categoria
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/models.Category'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Category'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Criar categoria
      tags:
      - Categories
  /categories/{id}:
    delete:
      parameters:
      - description: ID da categoria
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: Categoria excluída com sucesso
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Excluir categoria
      tags:
      - Categories
  /categories/{userId}:
    get:
      parameters:
      - description: ID do usuário
        in: path
        name: userId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Category'
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Listar categorias
      tags:
      - Categories
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"